carvaan/app/requirements.txt
pandas
python-fasthtml
ytmusicapi